requirements.txt
alembic==1.4.3
anyio==3.5.0
asgiref==3.5.2
attrs==21.4.0
Babel==2.10.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==7.1.2
coverage==6.3.2
fastapi==0.63.0
flake8==3.8.4
gunicorn==20.1.0
h11==0.11.0
idna==3.3
iniconfig==1.1.1
kafka-python==2.0.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
py-moneyed==2.0
pycodestyle==2.6.0
pydantic==1.7.3
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
pymongo==4.1.1
pymongo-inmemory==0.2.8
pyparsing==3.0.9
pytest==7.1.1
python-dateutil==2.8.1
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
requests==2.27.1
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.22
starlette==0.13.6
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
